+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/Implementation/EntityFrameworkDiagnosticListener.cs
@@ -181,7 +181,11 @@ internal sealed class EntityFrameworkDiagnosticListener : ListenerHandler
 
                         try
                         {
-                            if (command is IDbCommand typedCommand && this.options.Filter?.Invoke(typedCommand) == false)
+                            var dbContext = this.dbContextFetcher.Fetch(payload);
+                            var dbContextDatabase = this.dbContextDatabaseFetcher.Fetch(dbContext);
+                            var providerName = this.providerNameFetcher.Fetch(dbContextDatabase);
+
+                            if (command is IDbCommand typedCommand && this.options.Filter?.Invoke(providerName, typedCommand) == false)
                             {
                                 EntityFrameworkInstrumentationEventSource.Log.CommandIsFilteredOut(activity.OperationName);
                                 activity.IsAllDataRequested = false;